Rename stored_message.py to mapper.py
