+++ bot/action/extra/messages/__init__.py
@@ -6,7 +6,7 @@ from bot.action.core.action import Action
 from bot.action.core.command import UnderscoredCommandBuilder
 from bot.action.core.command.usagemessage import CommandUsageMessage
 from bot.action.extra.messages import analyzer
-from bot.action.extra.messages.stored_message import StoredMessageMapper
+from bot.action.extra.messages.mapper import StoredMessageMapper
 from bot.action.userinfo import UserStorageHandler
 from bot.action.util.format import UserFormatter, DateFormatter
 from bot.action.util.textformat import FormattedText
